% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{dashboard}
\alias{dashboard}
\title{Run a Zoonotic TB Shiny Dashboard}
\usage{
dashboard()
}
\value{
Starts a shiny dashboard
}
\description{
This functions runs a Zoonotic TB dashboard that has been built using other
package functionality. The dashboard can be used to explore the global burden of Zoonotic TB
interactively.
}
\examples{

## Only run the example if in an interative session
\dontrun{

## Run the TB dashboard
dashboard()
}
}
