% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{Demographic data}
\format{A data frame with 14,857 rows and 5 variables.
\describe{
\item{country}{Country name (factor)}
\item{country_code}{Country code (numeric)}
\item{year}{Year (numeric)}
\item{population}{Estimated population size (numeric)}
\item{prop_rural}{Proportion of population that is estimated to be rural (numeric)}
}}
\usage{
demographics
}
\description{
Country level demographic data from 1950 to the current day. Also includes the proportion of the population
that are rural. Data is sourced from the \href{http://www.fao.org/faostat/en/?#data/OA}{FAO}.
Data cleaning details can be found in \href{https://github.com/seabbs/EstZoonoticTB/blob/master/data-raw/demographics.R}{\\data-raw\\demographics.R}.
}
\keyword{datasets}
