% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstZoonoticTB.R
\docType{package}
\name{EstZoonoticTB-package}
\alias{EstZoonoticTB}
\alias{EstZoonoticTB-package}
\title{EstZoonoticTB: Explore, Visualise and Estimate the Global Zoonotic Tuberculosis Burden}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

'EstZoonoticTB' is an R package containing data relevant to global
        zoonotic tuberculosis (TB), tools for manipulating and visualising this data, and
        analysis aiming to improve country level estimates of zoonotic TB. Packaged 
        datasets include: cleaned data from a recent systematic review 
        (<https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4816377/>); data on the country 
        specific epidemiology of TB; demographic data (including data on rural
        populations); animal demographic data; and data on the presence of zoonotic TB 
        in animal populations (both domesticated and wild). Tooling includes functions
        for linking the built in datasets (built with the aim of accommodating
        external data sources), data mapping functions, and conveneance functions for
        manipulating the linked datasets. See the package vignettes for further details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.samabbott.co.uk/EstZoonoticTB}
  \item \url{https://github.com/seabbs/EstZoonoticTB}
  \item Report bugs at \url{https://github.com/seabbs/EstZoonoticTB/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (0000-0001-8057-8037)

}
\keyword{internal}
