% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_combined_data.R
\name{get_latest_combined_data}
\alias{get_latest_combined_data}
\title{Extract the Lastest Data for Each Variable of Interest}
\usage{
get_latest_combined_data(data)
}
\arguments{
\item{data}{Dataframe containing variables to be mapped. Must contain a \code{country_code} variable.
This must use iso3 standard country codes. See \code{\link[EstZoonoticTB]{link_data}} for an example dataframe.}
}
\value{
A dataframe containing the latest data on each variable of interest - along with a flag
indicating which year the data is from.
}
\description{
Extract the Lastest Data for Each Variable of Interest
}
\examples{

df <- link_data(verbose = FALSE)

get_latest_combined_data(df)

}
