% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_data.R
\name{get_latest_data}
\alias{get_latest_data}
\title{Extract the Latest Data for a Variable}
\usage{
get_latest_data(data, variable = NULL)
}
\arguments{
\item{data}{Dataframe containing variables to be mapped. Must contain a \code{country_code} variable.
This must use iso3 standard country codes. See \code{\link[EstZoonoticTB]{link_data}} for an example dataframe.}

\item{variable}{An unquoted variable name to select the most recent data for.}
}
\value{
A tibble containing the most recent estimates, by country, of the proportion of TB cases that are zoonotic
}
\description{
This conveniance function works with output from \code{[EstZoonoticTB]{link_data}} to extract
the most recent data from each country for a given variable
}
\examples{


## Extract linked data
df <- link_data(verbose = FALSE)

## Extract most recent study data
get_latest_data(df, tb_z_prop)
}
