% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data.R
\name{link_data}
\alias{link_data}
\title{Link Zoonotic TB Relevant Data}
\usage{
link_data(z_tb_humans = NULL, tb_humans = NULL, z_tb_animals = NULL,
  demo = NULL, animal_demo = NULL, verbose = TRUE)
}
\arguments{
\item{z_tb_humans}{Data on zoonotic TB in humans - formatted as \code{EstZoonoticTB::zoonotic_tb_humans}.
Defaults to \code{EstZoonoticTB::zoonotic_tb_humans}.}

\item{tb_humans}{Data on  TB in humans - formatted as \code{EstZoonoticTB::tb_data()}.
Defaults to \code{EstZoonoticTB::tb_data()}.}

\item{z_tb_animals}{Data on zoonotic TB in animals - formatted as \code{EstZoonoticTB::zoonotic_tb_animals}.
Defaults to \code{EstZoonoticTB::zoonotic_tb_humans}.}

\item{demo}{Data on demographics - formatted as \code{EstZoonoticTB::demographics}.
Defaults to \code{EstZoonoticTB::demographics}.}

\item{animal_demo}{Data on demographics in animals - formatted as \code{EstZoonoticTB::animal_demographics}.
Defaults to \code{EstZoonoticTB::animal_demographics}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose progress messages be shown.}
}
\value{
A dataframe of linked data containing all available global data relevant to zTB in humans.
}
\description{
Link data relevant to zoonotic TB data. Data is linking process
maximises preserved data so any analysis with this data may need additional exclusions
and/or munging.
}
\examples{


## Linked data relying on package defaults
df <- link_data()


df
}
