% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoonotic_tb_animals.R
\docType{data}
\name{zoonotic_tb_animals}
\alias{zoonotic_tb_animals}
\title{Status of zoonotic TB in animals}
\format{A data frame with 710 rows and 6 variables.
\describe{
\item{country}{Country name (factor)}
\item{country_code}{Country code (factor)}
\item{year}{Year (numeric)}
\item{half}{Year half (numeric)}
\item{animal}{Animal type (factor: "dom" = domesticated; "wild" = wild)}
\item{present}{Presence of zoonotic TB in animals (factor: present; limited; suspected; limited + suspected; not present; NA)}
}}
\usage{
zoonotic_tb_animals
}
\description{
Country level data on the status of zoonotic TB in animals (domesticated and wild) for 2018. Data is sourced from the \href{https://www.oie.int}{OIE}
via personal communication. Data cleaning details can be found in \href{https://github.com/seabbs/EstZoonoticTB/blob/master/data-raw/zoonotic_tb_animals.R}{/data-raw/zoonotic_tb_animals.R}.
}
\keyword{datasets}
