% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstZoonoticTB.R
\docType{package}
\name{EstZoonoticTB-package}
\alias{EstZoonoticTB}
\alias{EstZoonoticTB-package}
\title{EstZoonoticTB: Estimate and Visualise Zoonotic Tuberculosis}
\description{
Estimate and visualise zoonotic tuberculosis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.samabbott.co.uk/EstZoonoticTB}
  \item \url{https://github.com/seabbs/EstZoonoticTB}
  \item Report bugs at \url{https://github.com/seabbs/EstZoonoticTB/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (0000-0001-8057-8037)

}
\keyword{internal}
