% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_data.R
\name{tb_data}
\alias{tb_data}
\title{Tuberculosis data from the World Health Organization (WHO).}
\usage{
tb_data(inc_floor = NULL, inc_rate_floor = NULL, ...)
}
\arguments{
\item{inc_floor}{Numeric, defaults to NULL. What is the minimum incidence to keep in the data.}

\item{inc_rate_floor}{Numeric, defaults to NULL. What is the minimum incidence
rate (per 100,000) to keep in the data.}

\item{...}{Additional arguments to pass to \code{getTBinR::get_tb_burden}.}
}
\value{
A dataframe containing curated TB data from the WHO.
}
\description{
This function relies on \href{https://www.samabbott.co.uk/getTBinR/}{getTBinR} to source WHO tuberculosis (TB)
data. It then curates this data - extracting the country, region, year, TB incidence,
TB incidence rates (+ CI's), the proportion of cases with extra-pulmonary TB, and the proportion of cases with HIV (+ CI's).
Data prior to 2000 is dropped.
}
\examples{

## Get the data
tb_data()
}
\seealso{
get_tb_burden
}
