% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoonotic_tb_humans.R
\docType{data}
\name{zoonotic_tb_humans}
\alias{zoonotic_tb_humans}
\title{Status of zoonotic TB in humans}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 201 rows and 12 columns.}
\usage{
zoonotic_tb_humans
}
\description{
Country level data on the status of zoonotic TB in humans across multiple years. Data is sourced from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4816377/}{this systematic review}.
via personal communication. Data cleaning details can be found in \href{https://github.com/seabbs/EstZoonoticTB/blob/master/data-raw/zoonotic_tb_humans.R}{/data-raw/zoonotic_tb_humans.R}.
}
\keyword{datasets}
